+++ test/unit/pulled_search/unit_test_run3.sh
@@ -9,14 +9,15 @@ echo "Unit test:  pulled_search.py"
 /usr/bin/python3 ./test/unit/pulled_search/checks_dirs.py
 /usr/bin/python3 ./test/unit/pulled_search/cleanup_files.py
 /usr/bin/python3 ./test/unit/pulled_search/config_override.py
-/usr/bin/python3 ./test/unit/pulled_search/create_json.py
 /usr/bin/python3 ./test/unit/pulled_search/get_archive_files.py
 /usr/bin/python3 ./test/unit/pulled_search/help_message.py
 /usr/bin/python3 ./test/unit/pulled_search/insert_data.py
+/usr/bin/python3 ./test/unit/pulled_search/insert_mongo.py
 /usr/bin/python3 ./test/unit/pulled_search/load_processed.py
 /usr/bin/python3 ./test/unit/pulled_search/main.py
 /usr/bin/python3 ./test/unit/pulled_search/mvalidate_dirs.py
 /usr/bin/python3 ./test/unit/pulled_search/non_processed.py
+/usr/bin/python3 ./test/unit/pulled_search/parse_data.py
 /usr/bin/python3 ./test/unit/pulled_search/process_docid.py
 /usr/bin/python3 ./test/unit/pulled_search/process_failed.py
 /usr/bin/python3 ./test/unit/pulled_search/process_files.py
@@ -26,4 +27,3 @@ echo "Unit test:  pulled_search.py"
 /usr/bin/python3 ./test/unit/pulled_search/run_program.py
 /usr/bin/python3 ./test/unit/pulled_search/update_processed.py
 /usr/bin/python3 ./test/unit/pulled_search/validate_dirs.py
-/usr/bin/python3 ./test/unit/pulled_search/zgrep_search.py